Cloudinary Uploads
Within uploader we can now upload images to cloudinary.
The images are uploaded to the service but also details about
the image are stored within our database.
When we want to display the images we can just query the database for
the cloudinary url of the image we previously uploaded.
Error handling is still poor need to complete this before image panel
design and removing image functionality