Ability to add directly from the browser.
Using the cloudinary scripts provided from there blog we can now upload
images directly from the browser. What happens is that cloudinary adds a
hidden input when dragging photos to the upload area. The image is then
directly upload to cloudinary and we display a preview panel which
allows a user to see their photo and decided if they would like to
remove it. If the remove the image a request is then sent to cloudinary
to do so. If the user is happy with the image we send the data returned
from cloudinary to the server to keep a local copy of the image data e.g
urls etc.